--- abjadext/nauert/SerialJobHandler.py
@@ -1,23 +1,0 @@
-from .JobHandler import JobHandler
-
-
-class SerialJobHandler(JobHandler):
-    """
-    Serial job-handler.
-    """
-
-    ### CLASS VARIABLES ###
-
-    __slots__ = ()
-
-    ### SPECIAL METHODS ###
-
-    def __call__(self, jobs):
-        """
-        Calls serial job handler.
-
-        Returns ``jobs``.
-        """
-        for job in jobs:
-            job()
-        return jobs
